refactor(pet): 提取纯几何函数到 geometry 子模块
