+++ pinvou3-app/src-tauri/src/features/pet/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod detach;
+pub(crate) mod geometry;
 pub(crate) mod pet_window;
 mod platform;
 pub(crate) mod selected_pet;
